backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.9
alembic==1.15.2
pydantic-settings==2.9.1
python-jose[cryptography]==3.4.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
slowapi==0.1.9
python-multipart==0.0.20
email-validator==2.2.0
